+++ src/test/java/com/axibase/tsd/api/method/alert/AlertQueryTest.java
@@ -14,7 +14,7 @@ import java.util.Map;
 
 import static javax.ws.rs.core.Response.Status.OK;
 
-public class AlertQueryTest extends AlertMethod {
+public class AlertQueryTest extends AlertTest {
 
 
     /**
